fixed 7d18a4e from https://gitee.com/wuliubao/aafwk_standard/pulls/1237 IssueNo:#I4Y5M6 Description:codeDex告警Fopen无需判断 Sig: SIG_ApplicationFramework Feature or Bugfix: Feature Binary Source: No

+++ frameworks/kits/ability/native/src/datashare_stub.cpp
@@ -395,7 +395,7 @@ ErrCode DataShareStub::CmdExecuteBatch(MessageParcel &data, MessageParcel &reply
     }
 
     std::vector<std::shared_ptr<AppExecFwk::DataAbilityResult>> results = ExecuteBatch(operations);
-    int total = results.size();
+    int total = (int)(results.size());
     if (!reply.WriteInt32(total)) {
         HILOG_ERROR("DataShareStub::CmdExecuteBatchInner fail to WriteInt32 ret");
         return ERR_INVALID_VALUE;
