+++ interfaces/innerkits/form_manager/src/form_mgr_stub.cpp
@@ -33,7 +33,7 @@ void SplitString(const std::string &source, std::vector<std::string> &strings)
     }
     HILOG_DEBUG("the dump string split into %{public}d size", splitSize);
     for (size_t i = 0; i < splitSize; i++) {
-        int32_t start = LIMIT_PARCEL_SIZE * i;
+        size_t start = LIMIT_PARCEL_SIZE * i;
         strings.emplace_back(source.substr(start, LIMIT_PARCEL_SIZE));
     }
 }
